# backend/api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Velocity API"
  },
  "basePath": "/v1",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/projects": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Returns all of the projects",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "All of the projects",
            "schema": {
              "type": "object",
              "properties": {
                "total": {
                  "type": "integer",
                  "format": "long"
                },
                "result": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/ProjectResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Creates a new project",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ProjectRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successfully created",
            "schema": {
              "$ref": "#/definitions/ProjectResponse"
            }
          },
          "400": {
            "description": "Bad request. TODO: present validation errors"
          }
        }
      }
    },
    "/projects/{projectId}": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Returns a project given by its ID",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "ID of project to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "The project corresponding to the given ID",
            "schema": {
              "$ref": "#/definitions/ProjectResponse"
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "delete": {
        "tags": [
          "projects"
        ],
        "summary": "Deletes a project given by its ID",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "ID of project to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Succesfully deleted the project"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
      "flow": "implicit",
      "scopes": {
        "write:pets": "modify pets in your account",
        "read:pets": "read your pets"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "definitions": {
...
    "ProjectResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": "example-project"
        },
        "name": {
          "type": "string",
          "example": "Example Project"
        },
        "repository": {
          "type": "string",
          "example": "https://example-project.git"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "synchronising": {
          "type": "boolean",
          "example": false
        }
      }
    },
    "ProjectRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Example Project"
        },
        "repository": {
          "type": "string",
          "example": "https://example-project.git"
        },
        "key": {
          "type": "string",
          "example": ""
        }
      }
    },
...
  },
...
}